requirements.txt
psycopg2==2.7.3.2
Pillow==5.0.0
python-resize-image==1.1.11
reportlab==2.7
